#165 Directories refactor
