+++ Source/Cloud9/Contollers/Cloud9MouseController.h
@@ -26,9 +26,6 @@
 #include "CoreMinimal.h"
 #include "Cloud9ControllerComponent.h"
 
-#include "Cloud9/Tools/Macro/Common.h"
-#include "Cloud9/Tools/Macro/Logging.h"
-
 #include "Cloud9MouseController.generated.h"
 
 class ACloud9Character;
