+++ Source/Cloud9/Contollers/Cloud9PlayerController.h
@@ -26,9 +26,6 @@
 #include "CoreMinimal.h"
 #include "GameFramework/PlayerController.h"
 
-#include "Cloud9/Tools/Macro/Common.h"
-#include "Cloud9/Tools/Macro/Logging.h"
-
 #include "Cloud9PlayerController.generated.h"
 
 class UCloud9KeyboardController;
