+++ Source/Cloud9/Game/Cloud9DeveloperSettings.h
@@ -25,9 +25,6 @@
 
 #include "CoreMinimal.h"
 
-#include "Cloud9/Tools/Macro/Common.h"
-#include "Cloud9/Tools/Macro/Logging.h"
-
 #include "Cloud9DeveloperSettings.generated.h"
 
 USTRUCT(BlueprintType)
