+++ Source/Cloud9/Game/Cloud9GameState.cpp
@@ -1,13 +1,12 @@
 ﻿// Copyright (c) 2024 Alexei Gladkikh
 
-
 #include "Cloud9GameState.h"
-#include "Cloud9DeveloperSettings.h"
 
 #include "Cloud9/Tools/Macro/Common.h"
 #include "Cloud9/Tools/Macro/Logging.h"
 
 #include "Cloud9/Character/Cloud9Character.h"
+#include "Cloud9/Game/Cloud9DeveloperSettings.h"
 
 void ACloud9GameState::Tick(float DeltaSeconds)
 {
