+++ Source/Cloud9/Weapon/Classes/Cloud9WeaponGrenade.cpp
@@ -24,15 +24,15 @@
 #include "Cloud9WeaponGrenade.h"
 
 #include "DrawDebugHelpers.h"
+#include "PhysicsEngine/RadialForceComponent.h"
 
 #include "Cloud9/Tools/Extensions/TVariant.h"
 #include "Cloud9/Character/Cloud9Character.h"
 #include "Cloud9/Game/Cloud9DeveloperSettings.h"
 #include "Cloud9/Game/Cloud9GameInstance.h"
-#include "Cloud9/Game/Cloud9PlayerController.h"
+#include "Cloud9/Contollers/Cloud9PlayerController.h"
 #include "Cloud9/Weapon/Sounds/Cloud9SoundPlayer.h"
 #include "Cloud9/Weapon/Tables/WeaponTableGrenade.h"
-#include "PhysicsEngine/RadialForceComponent.h"
 
 const FName ACloud9WeaponGrenade::ExplosionComponentName = TEXT("ExplosionComponent");
 const FName ACloud9WeaponGrenade::DetonationEffectComponentName = TEXT("DetonationEffectComponent");
